Move test utility functions in a separate file
