+++ package.json
@@ -38,7 +38,7 @@
     "eslint-plugin-jest": "^22.2.2",
     "eslint-plugin-node": "^8.0.1",
     "eslint-plugin-prettier": "^3.0.1",
-    "jest": "^24.0.0",
+    "jest": "^24.6.0",
     "lerna": "^3.10.8",
     "prettier": "^1.16.3"
   },
