+++ packages/tools/jest.config.js
@@ -2,4 +2,5 @@ module.exports = {
   collectCoverage: true,
   collectCoverageFrom: ["packages/**/*.{js}", "!**/node_modules/**"],
   roots: ["packages/"],
+  testMatch: ["**/(*.)+\\.test.js"],
 };
